fixbug checkout, success page, cart, login/logout, session, and footer

+++ application/controllers/Checkout.php
@@ -12,35 +12,39 @@ class Checkout extends CI_Controller
 
 	public function index()
 	{
-		$data['customer'] = $this->db->get_where('customer', ['mail' => $this->session->userdata('mail')])->row_array();
-		$customer['user_name'] = $data['customer']['mail'];
-		$this->load->view('layout/header_page', $customer);
+		$this->load->view('layout/header_page');
 		$this->load->view('pages/checkout');
 		$this->load->view('layout/footer');
 	}
 
 	public function proses_order()
 	{
-		$data_order = array(
-			'orderDate' => date('Y-m-d'),
-			'customerID' => 1,
-			'notes' => "itu notes"
-		);
-		$id_order = $this->mOrder->add_order($data_order);
-
-		if ($cart = $this->cart->contents()) {
-			foreach ($cart as $item) {
-				$data_detail = array(
-					'orderID' => $id_order,
-					'productID' => $item['id'],
-					'productQty' => $item['qty'],
-					'price' => $item['price']
-				);
-				$proses = $this->mOrder->add_detail_order($data_detail);
+		if($this->session->userdata('status') == "login"){
+			$idCustomers = $this->session->userdata('id');
+			$notes = $this->input->post('orderNotes');
+			$data_order = array(
+				'orderDate' => date('Y-m-d'),
+				'customerID' => $idCustomers,
+				'notes' => $notes
+			);
+			$id_order = $this->mOrder->add_order($data_order);
+	
+			if ($cart = $this->cart->contents()) {
+				foreach ($cart as $item) {
+					$data_detail = array(
+						'orderID' => $id_order,
+						'productID' => $item['id'],
+						'productQty' => $item['qty'],
+						'price' => $item['price']
+					);
+					$proses = $this->mOrder->add_detail_order($data_detail);
+				}
 			}
+	
+			$this->cart->destroy();
+			redirect('success');
+		}else{
+			redirect('login');
 		}
-
-		$this->cart->destroy();
-		redirect('success');
 	}
 }
